Reduce false positives: ignore paths in CLI
Tests, .git, and more should be able to be ignored when running in CLI mode.
This will reduce runtime, and the number of false positives.

Note that the webhook excludes purely based on the final filename, whereas this will exclude based on any part of the path (ie: parent directory names, or the filename).